Last validation functions moved
